n1: We don't have the "new" Qcom GPS.
Not sure where this came from but it breaks horribly on the n1.

Strangely, no one has filed a JIRA bugreport on it or complained in
any of the threads, even though it makes Google Maps Navigation
nearly useless.

Specifically, this causes the GPS to provide a single data point
and then reset and relock when a SIM is inserted and the 3G radio is
on.  (Although there is a possibility it depends on the carrier's APN
configuration since the "new" QCOM GPS stuff is all AGPS-related.)

Change-Id: Ida68eb1a84fd373d1d589f36ad0f3372231cbfec