find7: remove cpufreq config from ramdisk
Adding another level of config in the ramdisk ist
just confusing and not needed.

There should only be kernel config and optional
user config. The kernel config should define
the absolute default for that device without the
need to pimp it in the ramdisk.

Also remove some unused stuff and move the
read ahead config to the correct place

Change-Id: I854901653293c4db5806612ff7edc5918d34f42e