Update patch set 6
Patch Set 6:

Working now.  There are some quirks that I don't like (such as THREE phases of LVM configuration shell scripts and the sdcard daemon for /data/media always running) but it's working on both LVM and non-LVM devices.

I'll spend some time explaining tonight and/or adding comments, but:
In some cases, there's stuff that really only belongs in one configuration being run for both.  This is because some things (like the sdcard daemon service for /data/media) are really hard to make conditional - but also do not interfere with the other configuration.

The /data/media sdcard daemon can't be set to "disabled" for example - it has to be started by init.  The fuse daemons for sdcard0 and sdcard1 CAN be set to "disabled" because vold starts them.

Similarly, directories are created for all configurations regardless of configuration to make lvm_symlinks simpler.  The exception is /storage/sdcard0 which is a directory (mount point in fact) in a non-LVM configuration but is a symlink in an LVM configuration.

Patch-set: 6
Label: Code-Review=0